Rename to SCBridgeWallet
